+++ modulegraph/find_modules.py
@@ -16,7 +16,7 @@ import warnings
 
 import modulegraph.modulegraph as modulegraph
 from modulegraph.modulegraph import Alias
-from modulegraph.util import imp_find_module, imp_walk
+from modulegraph.util import imp_find_module
 
 __all__ = [
     'find_modules', 'parse_mf_results'